requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
python-magic==0.4.27
Flask-Limiter==3.5.0
PyMySQL==1.1.0
cryptography==41.0.8
Werkzeug==3.0.1
Jinja2==3.1.2
bcrypt==4.1.2
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
